start refactor of lambda
